Added changes
Added changes

+++ packages/web-doctors/src/components/JuniorDoctors/JDCaseSheet/panels/MedicinePrescription.tsx
@@ -723,7 +723,7 @@ export const MedicinePrescription: React.FC = () => {
           : '';
       const dosageCount =
         medicine.medicineTimings.length > 0
-          ? parseInt(medicine.medicineDosage) * medicine.medicineTimings.length
+          ? parseFloat(medicine.medicineDosage) * medicine.medicineTimings.length
           : medicine.medicineDosage;
       return (
         <div key={index} className={classes.medicineBox}>
